P4ADEV-1276 moved service
