+++ src/test/java/it/gov/pagopa/payhub/auth/service/ValidateTokenServiceTest.java
@@ -3,7 +3,6 @@ package it.gov.pagopa.payhub.auth.service;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import it.gov.pagopa.payhub.auth.exception.custom.InvalidTokenException;
-import it.gov.pagopa.payhub.auth.service.exchange.AccessTokenBuilderService;
 import it.gov.pagopa.payhub.auth.utils.JWTValidator;
 import org.junit.jupiter.api.Assertions;
